refactor: move hifigan descriminators
